ml_container/requirements.txt
Flask
tensorflow
Werkzeug
flask_cors
pillow
keras